eightandhalf-backend/eightandhalf-recommend/requirements.txt
protobuf~=5.28.3
grpcio~=1.67.1
pika~=1.3.2
PyMySQL~=1.1.1
numpy~=1.22.2
pandas>2.0.0,<3.0.0
scipy>=1.10.1,<=1.13.1
joblib~=1.4.2
librosa~=0.10.2.post1
requests~=2.32.3
scikit-learn~=1.5.2